Renaming of the multiselect used by select list
